Reduced warning count

+++ include/utils/PacketUtility.h
@@ -117,7 +117,7 @@ struct is_back_inserter : std::false_type {};
         if (!packet_utility_v2::write(iterator, (uint16_t)value.size(), end)) {
             return false;
         }
-        for (int i = 0;i < value.size(); i++) {
+        for (size_t i = 0;i < value.size(); i++) {
             if (!packet_utility_v2::write(iterator, value[i], end)) {
                 return false;
             }
